Move nonsmall_diffs function from utils.py to several test_*.py files
